Place time window dropdown inside histogram container
